refactor: localise signature helper to test
