+++ src/routes/alerts/alerts.controller.spec.ts
@@ -2,6 +2,7 @@ import { INestApplication } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import * as request from 'supertest';
 import { faker } from '@faker-js/faker';
+import * as crypto from 'crypto';
 import { TestCacheModule } from '@/datasources/cache/__tests__/test.cache.module';
 import { TestNetworkModule } from '@/datasources/network/__tests__/test.network.module';
 import { TestLoggingModule } from '@/logging/__tests__/test.logging.module';
@@ -11,9 +12,30 @@ import { CacheModule } from '@/datasources/cache/cache.module';
 import { RequestScopedLoggingModule } from '@/logging/logging.module';
 import { NetworkModule } from '@/datasources/network/network.module';
 import { alertBuilder } from '@/routes/alerts/entities/__tests__/alerts.builder';
-import { fakeTenderlySignature } from '@/routes/alerts/entities/__tests__/fakeTenderlySignature';
 import { IConfigurationService } from '@/config/configuration.service.interface';
 import { Alert } from '@/routes/alerts/entities/alert.dto.entity';
+
+// The `x-tenderly-signature` header contains a cryptographic signature. The webhook request signature is
+// a HMAC SHA256 hash of concatenated signing secret, request payload, and timestamp, in this order.
+// @see https://github.com/Tenderly/tenderly-docs/blob/d836e99fcc22f141a155688128548505b9fcbf9c/alerts/configuring-alert-destinations/configuring-alert-destinations.md?plain=1#L74
+function fakeTenderlySignature(args: {
+  signingKey: string;
+  alert: Alert;
+  timestamp: string;
+}): string {
+  // Create a HMAC SHA256 hash using the signing key
+  const hmac = crypto.createHmac('sha256', args.signingKey);
+
+  // Update the hash with the request body using utf8
+  hmac.update(JSON.stringify(args.alert), 'utf8');
+
+  // Update the hash with the request timestamp
+  // Note: Tenderly timestamps are Go `time.Time` format, e.g.
+  // 2023-10-25 08:30:30.386157172 +0000 UTC m=+3512798.196320121
+  hmac.update(args.timestamp);
+
+  return hmac.digest('hex');
+}
 
 describe('Alerts (Unit)', () => {
   let app: INestApplication;
